Move InMemoriam components to separate folder
