+++ src/pages/in-memoriam.js
@@ -1,5 +1,5 @@
 import React from "react";
-import InMemoriamSection from "../components/IndexPage/InMemoriamSection";
+import InMemoriamSection from "../components/IndexPage/InMemoriam/InMemoriamSection";
 
 export default () => {
   return (
